requirements.txt
annotated-types==0.7.0
anyio==4.5.0
blinker==1.8.2
blobfile==3.0.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
faiss-cpu==1.8.0.post1
filelock==3.16.1
fire==0.6.0
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
fsspec==2024.9.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.25.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
llama_models==0.0.36
llama_stack==0.0.36
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
ollama==0.3.3
packaging==24.1
pandas==2.2.3
pillow==10.4.0
prompt_toolkit==3.0.48
pycryptodomex==3.20.0
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
rich==13.8.1
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
sentence-transformers==3.1.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
sympy==1.13.3
termcolor==2.4.0
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.19.1
torch==2.4.1
tqdm==4.66.5
transformers==4.44.2
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
wcwidth==0.2.13
Werkzeug==3.0.4
